Fixed your directory structure for better implementation in the future.
