+++ src/main/java/org/example/MainApplication.java
@@ -1,124 +1,21 @@
 package org.example;
 
-import java.util.ArrayList;
 import java.util.Scanner;
+import org.example.controllers.ArrayOperations;
+import org.example.controllers.SurfDataIO;
+import org.example.models.SurfForecast;
+import org.example.views.ConsoleMenu;
 
-// Interface for Data Import
-interface DataImport {
-    /**
-     * Imports data.
-     * @return An ArrayList containing the imported data.
-     */
-    ArrayList<String> importData();
-}
-
-
-/**
- * Implementation of DataImport for importing surf line data.
- */
-class SurfLineDataImport implements DataImport {
-    /**
-     * Imports fake surf line data for demonstration.
-     * @return An ArrayList containing the imported data.
-     */
-    public ArrayList<String> importData() {
-        // Fake data for demonstration
-        ArrayList<String> data = new ArrayList<>();
-        data.add("Fake Data 1");
-        data.add("Fake Data 2");
-        return data;
-    }
-    }
-
-// Base User class
-class User {
-    /**
-     * A shared method for all users.
-     * @return null
-     */
-    public Object sharedMethod() {
-        System.out.println("This is a shared method.");
-        return null;
-    }
-}
-
-// Admin class extending User
-class Admin extends User {
-    /**
-     * An admin-specific method.
-     * @return 0
-     */
-    public short adminMethod() {
-        System.out.println("This is an admin-specific method.");
-        return 0;
-    }
-}
-
-// Customer class extending User
-class Customer extends User {
-    /**
-     * A customer-specific method.
-     * @return 0
-     */
-    public short customerMethod() {
-        System.out.println("This is a customer-specific method.");
-        return 0;
-    }
-}
-
-// Class for Array Operations (Module 1 & 2)
-class ArrayOperations {
-    /**
-     * Displays the array operations menu.
-     */
-    public void showMenu() {
-        System.out.println("Array Operations Menu:");
-        System.out.println("1. Get element at index (1D array)");
-        System.out.println("2. Set element at index (1D array)");
-        System.out.println("8. Get element at index (2D array)");
-        System.out.println("9. Set element at index (2D array)");
-        System.out.println("Enter your choice for array operations:");
-    }
-    /**
-     * Gets the element at the given index in a 1D array.
-     * @param arr The array.
-     * @param index The index.
-     * @return The element at the given index.
-     */
-    public int getAtIndex(int[] arr, int index) {
-        return arr[index];
-    }
-
-    /**
-     * Sets the element at the given index in a 1D array.
-     * @param arr The array.
-     * @param index The index.
-     * @param value The value to set.
-     */
-    public void setAtIndex(int[] arr, int index, int value) {
-        arr[index] = value;
-    }
-    /**
-     * Gets the element at the given row and column in a 2D array.
-     * @param arr The 2D array.
-     * @param row The row.
-     * @param col The column.
-     * @return The element at the given row and column.
-     */
-    public int getAtIndex(int[][] arr, int row, int col) {
-        return arr[row][col];
-    }
-    /**
-     * Sets the element at the given row and column in a 2D array.
-     * @param arr The 2D array.
-     * @param row The row.
-     * @param col The column.
-     * @param value The value to set.
-     */
-    public void setAtIndex(int[][] arr, int row, int col, int value) {
-        arr[row][col] = value;
-    }
-}
+/*
+* Peer Review - Sean Standen
+* I pulled all of your classes and your interface out of this file to clean up
+* your project.  It will be much easier moving forward to find things on the fly.
+* I also organized your classes into directories depending on the type.
+* Directories:
+* /controllers - Application logic that performs actions.
+* /models - Classes that represent data types
+* /views - Classes that render views.
+* */
 
 // Main Application class (Module 4)
 public class MainApplication {
