+++ src/main/java/org/example/models/SurfForecast.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.models;
 
 import java.util.ArrayList;
